added user auth for business routes

+++ app/api/business_routes.py
@@ -2,9 +2,8 @@ from pydoc import describe
 from flask import Blueprint, request, jsonify
 from app.models import db, Business
 from ..forms.business_form import BusinessForm
-from ..forms.edit_business_form import EditBusinessForm
 from flask_login import current_user, login_required
-
+from .auth_routes import validation_errors_to_error_messages
 business_routes = Blueprint("businesses", __name__, url_prefix="/businesses")
 
 #TODO: Have not tested routes live yet
@@ -40,39 +39,45 @@ def create_business():
     db.session.commit()
     return jsonify(new_business.to_dict()), 200
   else:
-    return {'errors': form.errors}
+    return {'errors': validation_errors_to_error_messages(form.errors)}, 401
 
 #Edit Business
 @business_routes.route("/<int:business_id>", methods=["PUT"])
 @login_required
 def edit_business(business_id):
-  form = EditBusinessForm()
+  form = BusinessForm()
   form['csrf_token'].data = request.cookies['csrf_token']
   if form.validate_on_submit():
     business = Business.query.get(business_id)
-    business.address = form.address.data,
-    business.description = form.description.data,
-    business.url = form.url.data,
-    business.phone = form.phone.data,
-    business.state = form.state.data,
-    business.city = form.city.data,
-    business.zipcode = form.zipcode.data,
-    business.open_time = form.open_time.data,
-    business.close_time = form.close_time.data,
-    business.preview_image = form.preview_image.data
-    db.session.commit()
-    return jsonify(business.to_dict()), 200
+    if business.owner_id == current_user.id:
+      business.address = form.address.data,
+      business.description = form.description.data,
+      business.url = form.url.data,
+      business.phone = form.phone.data,
+      business.state = form.state.data,
+      business.city = form.city.data,
+      business.zipcode = form.zipcode.data,
+      business.open_time = form.open_time.data,
+      business.close_time = form.close_time.data,
+      business.preview_image = form.preview_image.data
+      db.session.commit()
+      return jsonify(business.to_dict()), 200
+    else:
+      return {'errors': 'Unauthorized'}, 401
   else:
-    return {'errors': form.errors}
+    return {'errors': validation_errors_to_error_messages(form.errors)}, 401
 
 #Delete Business
 @business_routes.route("/<int:business_id>", methods=["DELETE"])
 @login_required
 def delete_business(business_id):
   business = Business.query.filter(Business.id == business_id).first()
-  db.session.delete(business)
-  db.session.commit()
-  return jsonify({
-        "message": "Business successfully deleted",
-        "status-code": 200
-    }), 200
+  if business.owner_id == current_user.id:
+    db.session.delete(business)
+    db.session.commit()
+    return jsonify({
+          "message": "Business successfully deleted",
+          "status-code": 200
+      }), 200
+  else:
+    return {"errors": "Unauthorized"} , 401
